requirements.txt
django>=5.0
python-telegram-bot>=21.0
youtube-transcript-api>=0.6.3
ollama>=0.4
faiss-cpu>=1.7
tiktoken>=0.7
python-dotenv>=1.0
numpy>=1.26
